+++ src/js/utils/row.js
@@ -5,7 +5,7 @@ const convertTime = require('./convert-time.js')
 const row = {
   build: (track) => {
     const id = document.createElement('td')
-    id.append(button.play(track))
+    id.append(button.render(track))
 
     const name = document.createElement('td')
     name.innerText = track.name